[WIP] Import initial os_searchlight role.
Import the initial verison of the os_searchlight role.

This code is based on the os_gnocchi role and most likely has some
artefacts from the copied role that need to be removed.

Very much work in progress, but comments or suggestions are welcome.

This role currently does the following:
- Installs a jdk, elasticsearch and searchlight
- Starts the searchlight services (api and listener)
- Does an initial DB sync
- Listens for messages from other openstack services
- installs the openstack CLI plugin

Current issues / TODO list:

- Go through code removing any unneded variables
- Remove any other code only needed for os_gnocchi
- Resolve current keystone issues in CLI
- Tests - I believe I've fixed up the testing infrastructure.
- Ubuntu 14.04 install - (14.04 is going to be removed very soon! focus
  on 16.04, if that works forget 14.04)
- Bring role in line with current best practices
- linting of Stouts.elasticsearch role

Change-Id: I3ff72c9fbfc96a1f80f5cdd51a91f4d7ef638682