[WIP] Import initial os_searchlight role.
Import the initial verison of the os_searchlight role.

This code is based on the os_gnocchi role and most likely has some
artefacts from the copied role that need to be removed.

Very much work in progress, but comments or suggestions are welcome.

This role currently does the following:
- Installs a jdk, elasticsearch and searchlight
- Starts the searchlight services (api and listener)
- Does an initial DB sync
- Listens for messages from other openstack services
- installs the openstack CLI plugin

Current issues / TODO list:
- linting of Stouts.elasticsearch role

Change-Id: I3ff72c9fbfc96a1f80f5cdd51a91f4d7ef638682