Make unstable functional test non-voting
Given that the functional build for this repo
rarely passes and the repo itself is currently
unused, we can set the functional test to
non-voting in order to allow general pattern
changes to be implemented without delay.

This does mean that reviewers applying functional
changes will have to be disciplined about checking
the results, but I think that is OK in this case.

Depends-On: https://review.openstack.org/556253
Change-Id: I3f795e59f1b58862f2f196b2ac06d9891c3a17fa
(cherry picked from commit 00e1b9575e3786cec3f62676e9c6d270cf116398)