Only replace python when re-initializing the venv
We reinitialize the venv to ensure that the right version of
python is in the venv, but we do not want virtualenv to also
replace pip, setuptools and wheel so we tell it not to. If we
don't do this then virtualenv will install the latest available
version, which is not what we want.

Change-Id: I39e64f29e794075ef303d3e1a2c0c1a6a18937b6
Partial-Bug: #1764470
(cherry picked from commit a9561bbce96a0dc545f18aaa88be0988be2ac3ad)